test: split test_context_admission_ledger.py by concern (issue #4606)
Closes #4606

Decomposes the 2620-line test_context_admission_ledger.py into five
concern-scoped modules per the audit's finding (C9.25 - Oversized files):

- test_context_admission_ledger_recovery.py     (26 tests)
- test_context_admission_ledger_projection.py   (19 tests)
- test_context_admission_ledger_journal.py      (7 tests)
- test_context_admission_ledger_contention.py   (11 tests)
- test_context_admission_ledger_sticky_health.py (7 tests)
- _context_admission_ledger_helpers.py          (shared _authority +
                                                 _GOLDEN_JOURNAL)

Plus a new arch guard test ensuring no future refactor re-merges the
split (tests/arch/test_context_admission_ledger_split.py).

Total: 70 test functions preserved (matches the pre-split count), 0
behavior changes. test-source-map.json refreshed - the eight source
files that referenced the original mega-file now point at the five
split modules.
